# scheduler,router: Wait for router drain when stopping jobs