This is an attempt to prevent 503s during deployments by waiting for requests to finish to backends before stopping jobs.

/cc @titanous @josephglanville