@josephglanville 2016-09-25
Tracking requests rather than attempting to attack the connection management is a cool idea, connection management is a gnarly proposition in comparison.
I wonder if adding a mutex acquire to every request has visible performance overhead though.