@lmars 2016-10-04
@titanous @josephglanville @imjorge I have updated things as follows:

* Added a `--drain-backends` flag when adding routes which enables backend request tracking and up / down / drained events
* Updated the scheduler to watch for router backend events and only wait for drain events for jobs which previously were marked as up by the router (i.e. only if the route has backend draining enabled)
* Added `SchedulerSuite.TestGracefulShutdown` which makes some long-lived HTTP requests and makes sure they are drained before the scheduler stops the backend jobs

I also ran `wrk` for 30s and performed multiple deployments (`flynn env set` in a loop).

Without this change:

```
$ wrk -t10 -c10 -d30s http://http-test.1.localflynn.com
Running 30s test @ http://http-test.1.localflynn.com
  10 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.33ms    1.45ms  26.92ms   90.24%
    Req/Sec     0.94k   130.66     1.72k    67.33%
  280384 requests in 30.05s, 31.82MB read
  Non-2xx or 3xx responses: 107
Requests/sec:   9329.55
Transfer/sec:      1.06MB

# many errors in the router logs
$ flynn -a router log -r | grep -F 'lvl=eror' | wc -l
470

# mostly ECONNREFUSED and ECONNRESET
$ flynn -a router log -r | grep -F 'lvl=eror' | grep -F 'connection refused' | wc -l
362
$ flynn -a router log -r | grep -F 'lvl=eror' | grep -F 'connection reset by peer' | wc -l
88
```

With this change:

```
$ wrk -t10 -c10 -d30s http://http-test.1.localflynn.com
Running 30s test @ http://http-test.1.localflynn.com
  10 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.35ms    1.60ms  32.41ms   91.21%
    Req/Sec     0.94k   140.51     1.43k    64.60%
  281359 requests in 30.04s, 31.93MB read
Requests/sec:   9364.86
Transfer/sec:      1.06MB

# No errors in the logs
$ flynn -a router log -r | grep -F 'lvl=eror' | wc -l
0
```

This was for the test HTTP app (`test/apps/http`).

I also did the same for a backend with more realistic response times (random between 50ms and 2000ms) and I found the same result, just with much smaller throughput so less errors (increasing the threads / connections leads to port exhaustion pretty quickly).