@titanous 2016-10-05
Is there a downside to defaulting the `--drain-backends` flag to true?