@lmars 2016-10-09
@titanous comments addressed